requirements.txt
cupy-cuda11x
scipy
numpy
sympy

